Ensure archetype comparison tests run for child modules
It appears that the archetype plugin does not do comparisons for nested archetypes so we must split the tests into two. A *check* test which does a single layer generate/comparison and a *build* test which uses the project archetype to make a top-level build and then builds the module generated by the archetype being tested.

Signed-off-by: Tim Ward <tim.ward@paremus.com>