Try using built in mvn as recommended by AppVeyor support
According to the AppVeyor support team the source of our slow builds is the maven wrapper that we use. This leads me to believe that AppVeyor have some special settings in their Maven installation which we lose by using the wrapper. This commit switches to using the AppVeyor installed mvn command

Signed-off-by: Tim Ward <tim.ward@paremus.com>